Classes : Priest + Classes : Pyromaniac + UI : ClassInGame
Priest :
- Skill cost
- cooldown
Pyromaniac
- Skill cost
- Cooldown
ClassInGame :
- update fillbar special type
- update fillbar health

TESTS :
- Change leader with right click
- Use ability with spacebar

+++ Assets/Classes/Scripts/DW_Class.cs
@@ -47,5 +47,5 @@ public class DW_Class : ScriptableObject
 
     public specialSourceType specialSourceType;
     [Range(0, 100)]
-    public int specialSourceAmount;
+    public float specialSourceAmount;
 }
